Some refactoring and returnign of data.
- Remove Set Name API as its no longer needed.
- Move Player into the player types folder.
- Return Lobby top 50 chats when someone joins so they have the chats to start with.
