+++ pkg/models/lobby/types.go
@@ -1,19 +1,14 @@
 package lobby
 
-type Player struct {
-	LobbyId      string `dynamodbav:"LobbyId" json:"-"`
-	SessionId    string `dynamodbav:"SessionId" json:"-"`
-	ConnectionId string `dynamodbav:"ConnectionId" json:"-"`
-	Id           string `dynamodbav:"Id" json:"id"`
-	Name         string `dynamodbav:"Name" json:"name"`
-	Points       int32  `dynamodbav:"Points" json:"points"`
-	IsAdmin      bool   `dynamodbav:"IsAdmin" json:"isAdmin"`
-	IsOnline     bool   `dynamodbav:"IsOnline" json:"isOnline"`
-}
+import (
+	"backend/pkg/models/chat"
+	"backend/pkg/models/player"
+)
 
 type Details struct {
-	Players []Player `json:"players"`
-	LobbyId string   `json:"lobbyId"`
+	Chats   []chat.Chat     `json:"chats"`
+	Players []player.Player `json:"players"`
+	LobbyId string          `json:"lobbyId"`
 }
 
 type Lobby struct {
